refactor: password만을 위한 훅 생성 및 공통 훅으로 이동 #293
- 관련 상수들도 common 폴더로 이동
