+++ frontend/__tests__/useUserIdInput.test.ts
@@ -2,7 +2,7 @@ import { renderHook, act } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
 import { ERROR_MESSAGE } from '../src/pages/signup/constants/errorMessage';
-import { ID } from '../src/pages/signup/constants/id';
+import { ID } from '../src/common/constants/id';
 import useUserIdInput from '../src/pages/signup/hooks/useUserIdInput';
 
 const createFakeChangeEvent = (
