+++ frontend/src/common/constants/password.ts
